% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dailyschedule.R
\name{DailySchedule}
\alias{DailySchedule}
\title{Return a table of game-ids and home vs away info for a given date}
\usage{
DailySchedule(date, API_KEY)
}
\arguments{
\item{date}{A Date formatted as YYYY/MM/DD}

\item{API_KEY}{A Key}
}
\value{
A table of the daily schedule
}
\description{
Return a table of game-ids and home vs away info for a given date
}
