% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goallocations.R
\name{GoalLocations}
\alias{GoalLocations}
\title{Return an interactive coordinate plane with locations of goals for a given game}
\arguments{
\item{game_id}{A Unique ID}

\item{API_KEY}{A Key}
}
\value{
A plane with points plotting x and y coordinates of a goal
}
\description{
Return an interactive coordinate plane with locations of goals for a given game
}
