% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outcomes.R
\name{Outcome}
\alias{Outcome}
\title{Return table of game outcomes}
\usage{
Outcome(game_id, API_KEY)
}
\arguments{
\item{game_id}{A Unique ID}

\item{API_KEY}{A Key}
}
\value{
A table of the outcomes of a game
}
\description{
Return table of game outcomes
}
