% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yearlyrankings.R
\name{YearlyRankings}
\alias{YearlyRankings}
\title{Return table of the yearly rankings, divided by conference and division for a given year}
\usage{
YearlyRankings(season_year, API_KEY)
}
\arguments{
\item{season_year}{A Year}

\item{API_KEY}{A Key}
}
\value{
A table of the outcomes of a game
}
\description{
Return table of the yearly rankings, divided by conference and division for a given year
}
\section{Functions}{
\itemize{
\item \code{YearlyRankings}: returns table
}}

